+++ src/Controller/StatusesController.php
@@ -11,15 +11,33 @@ use App\Controller\AppController;
 class StatusesController extends AppController {
 
 /**
+ * Initialize method
+ *
+ * @return void
+ */
+	public function initialize() {
+		parent::initialize();
+		$this->loadComponent('RequestHandler');
+		$this->response->header('Access-Control-Allow-Origin', '*');
+		$this->response->header('Access-Control-Allow-Methods', 'POST, GET, OPTIONS, PUT, DELETE');
+	}
+
+/**
  * Index method
  *
  * @return void
  */
 	public function index() {
-		$this->paginate = [
-			'contain' => ['Creators', 'Modifiers']
-		];
-		$this->set('statuses', $this->paginate($this->Statuses));
+		if ($this->request->params['_ext']) {
+			$conditions = [
+			//	'fields' => ['Statuses.id', 'Statuses.name']
+			];
+			$this->set('statuses', $this->Statuses->find('all', $conditions));
+		}
+		else {
+			$this->set('statuses', $this->Statuses->find('all'));
+		}
+		$this->set('_serialize', ['statuses']);
 	}
 
 /**
@@ -30,10 +48,17 @@ class StatusesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function view($id = null) {
-		$status = $this->Statuses->get($id, [
-			'contain' => ['Creators', 'Modifiers', 'Orders']
-		]);
-		$this->set('status', $status);
+		if ($this->request->params['_ext']) {
+			$status = $this->Statuses->get($id);
+			$this->set('status', $status);
+			$this->set('_serialize', ['status']);
+		} else {
+			$status = $this->Statuses->get($id, [
+				'contain' => ['Creators', 'Modifiers']
+			]);
+			$this->set('status', $status);
+			$this->set('_serialize', ['status']);
+		}
 	}
 
 /**
@@ -42,18 +67,33 @@ class StatusesController extends AppController {
  * @return void
  */
 	public function add() {
-		$status = $this->Statuses->newEntity($this->request->data);
-		if ($this->request->is('post')) {
-			if ($this->Statuses->save($status)) {
-				$this->Flash->success('The status has been saved.');
-				return $this->redirect(['action' => 'index']);
+		if ($this->request->params['_ext']) {
+			$status = $this->Statuses->newEntity($this->request->data);
+			if ($this->Statuses->save($status, ['validate' => false])) {
+				$message = 'Saved';
 			} else {
-				$this->Flash->error('The status could not be saved. Please, try again.');
+				$message = 'Error';
 			}
+			$this->set([
+				'data' => $this->request->data,
+				'message' => $message,
+				'status' => $status,
+				'_serialize' => ['message', 'status', 'data']
+			]);
+		} else {
+			$status = $this->Statuses->newEntity($this->request->data);
+			if ($this->request->is('post')) {
+				if ($this->Statuses->save($status)) {
+					$this->Flash->success('The status has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The status could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Statuses->Creators->find('list');
+			$modifiers = $this->Statuses->Modifiers->find('list');
+			$this->set(compact('status', 'creators', 'modifiers'));
 		}
-		$creators = $this->Statuses->Creators->find('list');
-		$modifiers = $this->Statuses->Modifiers->find('list');
-		$this->set(compact('status', 'creators', 'modifiers'));
 	}
 
 /**
@@ -64,21 +104,40 @@ class StatusesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function edit($id = null) {
-		$status = $this->Statuses->get($id, [
-			'contain' => []
-		]);
-		if ($this->request->is(['patch', 'post', 'put'])) {
-			$status = $this->Statuses->patchEntity($status, $this->request->data);
-			if ($this->Statuses->save($status)) {
-				$this->Flash->success('The status has been saved.');
-				return $this->redirect(['action' => 'index']);
-			} else {
-				$this->Flash->error('The status could not be saved. Please, try again.');
+		if ($this->request->params['_ext']) {
+			$status = $this->Statuses->get($id);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$status = $this->Statuses->patchEntity($status, $this->request->data);
+				if ($this->Statuses->save($status, ['validate' => false])) {
+					$message = 'Saved';
+				} else {
+					$message = 'Error';
+				}
 			}
+			$this->set([
+				'status' => $status,
+				'message' => $message,
+				'data' => $this->request->data,
+				'_serialize' => ['message','status','data']
+			]);
 		}
-		$creators = $this->Statuses->Creators->find('list');
-		$modifiers = $this->Statuses->Modifiers->find('list');
-		$this->set(compact('status', 'creators', 'modifiers'));
+		else {
+			$status = $this->Statuses->get($id, [
+				'contain' => []
+			]);
+			if ($this->request->is(['patch', 'post', 'put'])) {
+				$status = $this->Statuses->patchEntity($status, $this->request->data);
+				if ($this->Statuses->save($status)) {
+					$this->Flash->success('The status has been saved.');
+					return $this->redirect(['action' => 'index']);
+				} else {
+					$this->Flash->error('The status could not be saved. Please, try again.');
+				}
+			}
+			$creators = $this->Statuses->Creators->find('list');
+			$modifiers = $this->Statuses->Modifiers->find('list');
+			$this->set(compact('status', 'creators', 'modifiers'));
+		}
 	}
 
 /**
@@ -89,13 +148,26 @@ class StatusesController extends AppController {
  * @throws \Cake\Network\Exception\NotFoundException
  */
 	public function delete($id = null) {
-		$status = $this->Statuses->get($id);
-		$this->request->allowMethod(['post', 'delete']);
-		if ($this->Statuses->delete($status)) {
-			$this->Flash->success('The status has been deleted.');
-		} else {
-			$this->Flash->error('The status could not be deleted. Please, try again.');
+		if ($this->request->params['_ext']) {
+			$status = $this->Statuses->get($id);
+			$message = 'Deleted';
+			if (!$this->Statuses->delete($status)) {
+				$message = 'Error';
+			}
+			$this->set([
+				'message' => $message,
+				'_serialize' => ['message']
+			]);
 		}
-		return $this->redirect(['action' => 'index']);
+		else {		
+			$status = $this->Statuses->get($id);
+			$this->request->allowMethod(['post', 'delete']);
+			if ($this->Statuses->delete($status)) {
+				$this->Flash->success('The status has been deleted.');
+			} else {
+				$this->Flash->error('The status could not be deleted. Please, try again.');
+			}
+			return $this->redirect(['action' => 'index']);
+		}
 	}
 }
